Refactored some AdapterView specific methods into the BaseAdapterConverter.
