+++ UniversalAdapter/src/main/java/com/raizlabs/android/universaladapter/widget/adapters/converter/UniversalAdapter.java
@@ -8,7 +8,6 @@ import com.raizlabs.android.coreutils.threading.ThreadingUtils;
 import com.raizlabs.android.coreutils.util.observable.lists.ListObserver;
 import com.raizlabs.android.coreutils.util.observable.lists.ListObserverListener;
 import com.raizlabs.android.coreutils.util.observable.lists.SimpleListObserver;
-import com.raizlabs.android.universaladapter.widget.adapters.UniversalAdapterUtils;
 import com.raizlabs.android.universaladapter.widget.adapters.ViewHolder;
 
 public abstract class UniversalAdapter<Item, Holder extends ViewHolder> {
@@ -132,26 +131,7 @@ public abstract class UniversalAdapter<Item, Holder extends ViewHolder> {
 
     public abstract int getCount();
 
-    public abstract Object getItem(int position);
-
     public abstract Item get(int position);
-
-    public View getView(int position, View convertView, ViewGroup parent) {
-        Holder viewHolder = null;
-        if (convertView != null) {
-            viewHolder = getViewHolder(convertView);
-        }
-
-        if (viewHolder == null) {
-            int viewType = getItemViewType(position);
-            viewHolder = createViewHolder(parent, viewType);
-            setViewHolder(viewHolder.itemView, viewHolder);
-        }
-
-        bindViewHolder(viewHolder, position);
-
-        return viewHolder.itemView;
-    }
 
     public boolean areAllItemsEnabled() {
         return true;
@@ -171,38 +151,6 @@ public abstract class UniversalAdapter<Item, Holder extends ViewHolder> {
 
     public boolean hasStableIds() {
         return false;
-    }
-
-    public View getDropDownView(int position, View convertView, ViewGroup parent) {
-        Holder viewHolder = null;
-        if (convertView != null) {
-            viewHolder = getViewHolder(convertView);
-        }
-
-        if (viewHolder == null) {
-            int viewType = getItemViewType(position);
-            viewHolder = createDropDownViewHolder(parent, viewType);
-            setViewHolder(viewHolder.itemView, viewHolder);
-        }
-
-        bindDropDownViewHolder(viewHolder, position);
-
-        return viewHolder.itemView;
-    }
-
-    @SuppressWarnings("unchecked")
-    protected Holder getViewHolder(View view) {
-        try {
-            return (Holder) UniversalAdapterUtils.getViewHolder(view);
-        } catch (ClassCastException ex) {
-            // Don't care. Just don't crash. We'll just ignore convertView.
-        }
-
-        return null;
-    }
-
-    protected void setViewHolder(View view, Holder holder) {
-        UniversalAdapterUtils.setViewHolder(view, holder);
     }
 
     public Holder createViewHolder(ViewGroup parent, int itemType) {
